refactor swapstore
