+++ swap/store/domain/errors.go
@@ -1,0 +1,5 @@
+package swapdomain
+
+import "errors"
+
+var ErrNotFound = errors.New("not found")
